Refactor blogs folder structure
