+++ src/app/blog/blog.module.ts
@@ -4,11 +4,13 @@ import { CommonModule } from '@angular/common';
 import { BlogComponent } from './blog.component';
 import { SharedModule } from '../shared/shared.module';
 import { BlogRoutingModule } from './blog-routing.module';
-import { BlogsService } from '../shared/services/blogs/blogs.service';
+import { BlogsService } from './blogs.service';
+import { TopBlogsComponent } from './top-blogs/top-blogs.component';
 
 @NgModule({
-  declarations: [BlogComponent],
+  declarations: [BlogComponent, TopBlogsComponent],
   imports: [CommonModule, SharedModule, BlogRoutingModule],
+  exports: [TopBlogsComponent],
   providers: [BlogsService]
 })
 export class BlogModule {}
