+++ src/app/blog/top-blogs/top-blogs.component.ts
@@ -3,8 +3,8 @@ import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { Blog } from '../../shared/models/blog.model';
-import { BlogsService } from '../../shared/services/blogs/blogs.service';
-import { CloudFunctions } from '../../shared/services/data/cloud-functions';
+import { BlogsService } from '../blogs.service';
+import { CloudFunctions } from '../../shared/models/cloud-functions.constant';
 import { DataOrder } from '../../shared/models/data-type.model';
 
 const MAX_TOP_BLOGS_LENGTH = 3;
