+++ src/app/get-involved/country-supporters.service.ts
@@ -2,13 +2,14 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { AngularFirestore } from '@angular/fire/firestore';
 
-import { DataService } from '../shared/services/data/data.service';
+import { DataService } from '../shared/models/data-service.model';
 import { CountrySupporters } from '../shared/models/country-supporters.model';
-import { FirebaseCollections } from '../shared/services/data/collections';
+import { FirebaseCollections } from '../shared/models/collections.constant';
 
 @Injectable()
 export class CountrySupportersService extends DataService<CountrySupporters> {
   constructor(http: HttpClient, db: AngularFirestore) {
+    console.log('country')
     super(http, db, FirebaseCollections.countrySupporters);
   }
 }
