+++ src/app/home/home.module.ts
@@ -2,16 +2,16 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
-import { BlogsService } from '../shared/services/blogs/blogs.service';
+import { BlogsService } from '../blog/blogs.service';
 import { SharedModule } from '../shared/shared.module';
+import { BlogModule } from '../blog/blog.module';
+import { GetInvolvedModule } from '../get-involved/get-involved.module';
 import { HomeComponent } from './home.component';
 import { BannerComponent } from './banner/banner.component';
 import { ActivityIndicatorsComponent } from './activity-indicators/activity-indicators.component';
 import { TestimonialsComponent } from './testimonials/testimonials.component';
-import { GetInvolvedModule } from '../get-involved/get-involved.module';
 import { ActivityIndicatorsService } from './activity-indicators/activity-indicators.service';
 import { SponsorsComponent } from './sponsors/sponsors.component';
-import { TopBlogsComponent } from './top-blogs/top-blogs.component';
 import { SponsorsService } from './sponsors/sponsors.service';
 
 @NgModule({
@@ -21,9 +21,14 @@ import { SponsorsService } from './sponsors/sponsors.service';
     ActivityIndicatorsComponent,
     TestimonialsComponent,
     SponsorsComponent,
-    TopBlogsComponent
   ],
-  imports: [CommonModule, RouterModule, SharedModule, GetInvolvedModule],
+  imports: [
+    CommonModule,
+    RouterModule,
+    SharedModule,
+    GetInvolvedModule,
+    BlogModule
+  ],
   providers: [ActivityIndicatorsService, BlogsService, SponsorsService]
 })
 export class HomeModule {}
